refactor names for components
