+++ src/pagesMinin/Tables/CardsTable/CardsTable.module.scss
