+++ src/pagesMinin/Tables/DecksTable/decksTable.module.scss
