+++ src/pagesMinin/Tables/tableSingleRow.module.scss
