+++ src/pagesMinin/cardsPage.module.scss
