+++ src/pagesMinin/componentsMinin/Page/Page.tsx
@@ -7,7 +7,7 @@ import s from './page.module.scss'
 type Props = {
   mt?: CSSProperties['marginTop']
 } & ComponentPropsWithoutRef<'div'>
-export const PageMinin = ({ className, mt = '33px', style, ...rest }: Props) => {
+export const Page = ({ className, mt = '33px', style, ...rest }: Props) => {
   const classes = clsx(className, s.container)
   const styles: CSSProperties = { marginTop: mt, ...style }
