+++ src/pagesMinin/componentsMinin/Page/page.module.scss
